+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added override.conf for nm-wait-online-initrd.service with dracut.
 - Added userdocs for `wwctl node import` from yaml/csv.
 - Added uid, gid, and permissions to OverlayFile in REST API. #1925
+- (preview) Support provisioning to local storage with `wwctl <node|partition> set --root=/path/to/disk`
+- (preview) Support disk provisioning during dracut boot stage with Ignition, sfdisk, mkfs, and mkswap
+- Added systemd.mount and systemd.swap
+- Support configuring Ignition with resources
 
 ### Fixed